--- modules/apps/analytics/analytics-dxp-entity-impl/src/main/java/com/liferay/analytics/dxp/entity/internal/exporter/odata/entity/AnalyticsDXPEntityEntityModel.java
@@ -1,54 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.analytics.dxp.entity.internal.exporter.odata.entity;
-
-import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.odata.entity.DateTimeEntityField;
-import com.liferay.portal.odata.entity.EntityField;
-import com.liferay.portal.odata.entity.EntityModel;
-import com.liferay.portal.odata.entity.IntegerEntityField;
-
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-/**
- * @author Marcos Martins
- */
-public class AnalyticsDXPEntityEntityModel implements EntityModel {
-
-	public AnalyticsDXPEntityEntityModel() {
-		_entityFieldsMap = Stream.of(
-			new DateTimeEntityField(
-				"modifiedDate",
-				locale -> Field.getSortableFieldName(Field.MODIFIED_DATE),
-				locale -> Field.MODIFIED_DATE),
-			new IntegerEntityField(
-				Field.getSortableFieldName(Field.MODIFIED_DATE),
-				locale -> Field.getSortableFieldName(Field.MODIFIED_DATE))
-		).collect(
-			Collectors.toMap(EntityField::getName, Function.identity())
-		);
-	}
-
-	@Override
-	public Map<String, EntityField> getEntityFieldsMap() {
-		return _entityFieldsMap;
-	}
-
-	private final Map<String, EntityField> _entityFieldsMap;
-
-}
